chore: decrease inventory canary alarm severity (#733)
On the public Construct Hub we've received alarms about the inventory canary timing out due to the large number of packages it is trying to analyze. This inventory canary isn't critical to the Construct Hub's customer-facing behavior since none of the metrics are directly tied to other alarms, so I've changed the alarm to low severity.

I also increased the timeout / schedule rate of the canary to resolve the actual timeout issue we are having, and created the issue #732 for tracking a long-term fix.