chore: increase reprocessing rate (#734)
The changes introduced in https://github.com/cdklabs/construct-hub/pull/731 over-estimated the amount of backoff needed, causing the updated reprocessing job to run at an extremely low rate (never exceeding 20 concurrent tasks at a time) -- so it would end up taking over a day to finish.

The math was calculated right, but it didn't take account for the fact that the jobs take several minutes to be dispatched by the state machine. That is, when a state machine takes 5 minutes to execute 1000 tasks with the current setup, it isn't actually 1000 tasks running in parallel for 5 minutes each, but 1000 tasks running for around a minute each, staggered over 4 minutes. (just example numbers here)

It's hard to be more precise here without over-modeling the problem.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*