fix: inventory canary reporting packages as unknown (#752)
Our package canary was reporting most packages as unknown or missing because of a bug where we were matching against keys incorrectly. This PR fixes the bug, and cleans up the rest of the canary code to properly handle the fact that we are generating both json and markdown documents, but do not want to double count the number of identified packages (nor count either type of file as "unknown").

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*