fix: transliterator fails when reprocessing an uninstallable package (#760)
Fixes #759

In some cases, we noticed messages appearing in our Orchestration DLQ with the message "Forbidden: null". This message suggests a permission error. We identified the most likely cause was the transliterator ECS task did not have the correct permission to read "uninstallable" object markers that we assign to packages that fail to install with `npm install`. We were currently granting write and delete permissions, but not read permissions.

I would add a unit test for our transliterator to prevent a regression for this path in the code, but the cause of this bug is a permission error so it wouldn't really be caught by any test we write unless it was a full-on integration test, which is a bit out of the scope of this PR. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*