fix: IAM grants are not correctly added to VPC endpoint (#766)
The changes added in https://github.com/cdklabs/construct-hub/pull/760 had the unintended effect of actually _decreasing_ the number of files that the documentation transliterator was being given access to. The reason for this is that the code is using a Proxy object to duplicate IAM grants from the s3 bucket policy to the VPC endpoint policy, but it is only configured to support a handful of `grant` methods. This did log a warning, but it was not noticed in the test output.

I fixed the bug by using the `grantRead` and `grantWrite` permissions separately, and updating this Proxy code to throw the error instead of logging it to prevent this kind of error occur in the future.

This bug took a decent amount of time to discover because my dev stack was created with the default options (without a VPC), whereas the public Construct Hub is deployed with a VPC, so I wasn't able to reproduce the error locally.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*