fix: decrease severity of DLQ alarm (#774)
We've seen a number of occurrences where the DLQ alarm gets triggered as a "high severity" event for us, but in practice there is little to no user-facing impact (at worst, the user is seeing documentation that has been generated with an outdated version of `jsii-docgen` or `jsii-rosetta` etc). If new packages are not getting properly transliterated, this will generally be caught by our package canary. In lieu of this, this PR decreases the severity of the alarm. Fixes #767.

In addition, we noticed a very small number of ECS tasks (around 1 or 2, never the same ones) fail during reprocessing because of transient timeout errors. The cause doesn't seem to be that they are actually taking a long time to run, but that the tasks are taking a bit longer than 5 minutes to provision so it doesn't even get to send the first heartbeat back to the caller. This PR increases the heartbeat timeout and the number of retries we apply in this scenario to decrease the probability these items get sent to the DLQ.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*