chore: mitigate alarm verbosity
In the last few weeks we have regularly observed a large gap between the
npmjs replica database (replicate.npmjs.com/registry) and the registry's
primary endpoint (registry.npmjs.org). When this happens, the SLA breach
alarm triggers despite ConstructHub cannot possible have indexed new
packages (as these are not visible on the replica just yet).

To reduce inactionnable alarms, this adds a process to measure the lag
that is observed between the primary and replica, and suppresses the
SLA alarm in case the delta is found to be too large to allow the SLA to
be met.

Additionally, the SLA alarm triggering is not necessarily indicative of
an operational issue with ConstructHub, and will not be immediately
customer-affecting. As a consequence, it was downgraded from high- to
low-severity.