fix(npmjs): canary alarms when replicate.npmjs.com is down (#899)
Sometimes, the NpmJS canary errors because replicate.npmjs.com is down, causing an alarm to be triggered. There's no use failing and alarming since there's nothing that that can be done by the operator. So this PR catches the error, and opts to simply not report the NPM replica lag metric when it's unavailable.

The Npmjs/Canary/SLA-Breached alarm, which depends on this metric to determine if replicate.npmjs.com is down (and disable the alarm accordingly), will now fill in values based on what the last reported lag-time metric was, instead of filling in empty values with 0.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*