fix: NpmJs source canary fails when npm servers are impaired
We are seeing semi-frequent occurrences of npm servers being impaired
and returning HTTP Gateway errors (HTTP 502, HTTP 504). There is nothing
we can do but wait for these to resolve, so we should not be failing and
alarming unless the situation persists for an extended amount of time.

This change catches HTTP errors and "ignores" HTTP 502 and HTTP 504
errors, while emitting a metric to reflect those errors being
encountered. A new low-severity alarm will fire if we continue to see
HTTP 502 and HTTP 504 errors on every single run for more than 60
minutes.