fix: NpmJs source canary fails when npm servers are impaired (#906)
We are seeing semi-frequent occurrences of npm servers being impaired
and returning HTTP Gateway errors (HTTP 502, HTTP 504). There is nothing
we can do but wait for these to resolve, so we should not be failing and
alarming unless the situation persists for an extended amount of time.

This change catches HTTP errors and "ignores" HTTP 502 and HTTP 504
errors, while emitting a metric to reflect those errors being
encountered. A new low-severity alarm will fire if we continue to see
HTTP 502 and HTTP 504 errors on every single run for more than 60
minutes.

I chose not to apply a back-off & retry policy as those issues are
usually caused by servers being overloaded and the best strategy here
is to let these servers cool off, not pile up more load on them.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*