chore: increase timeout of NpmJs canary
When the npm replica is overloaded and we get HTTP 504 (Gateway Timeout),
these error responses can take upstream of 30 seconds to be returned, which
makes it likely that the function times out if it's operating on a 1 minute
timeout.

Reducing the request timeout is difficult as this would introduce a new
error class to handler (timeouts), and the handling of these would be
awkward (should be similar to gateway timeout, so why not wait for this?).