chore: set HTTP request timeouts in NpmJs Canary (#908)
When the npm replica is overloaded and we get HTTP 504 (Gateway Timeout),
these error responses can take upstream of 30 seconds to be returned, which
makes it likely that the function times out if it's operating on a 1 minute
timeout.

Setting socket timeouts on the sockets for requests sent to replicate.npmjs.com
to 5 seconds in order to alleviate this problem. Those get recorded as if they
were gateway timeout errors.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*