fix: ingestion lambda incorrectly compares assembly filenames (#943)
Previously, the callback function in the ingestion lambda was comparing a file name to a file path, resulting in comparisons like `'package/.jsii.gz' === '.jsii.gz'`. This PR changes that behavior to validate only that the filename, `.jsii.gz` is what we expect. That is all the information we have in the redirect schema.

In addition, adds a test that honestly should have been there in the first place, which ensures that the ingestion lambda can handle compressed assemblies.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*