fix: raise timeout of Version Tracker Lambda
Version tracker regularly runs out of the 1 minute timeout. Set it to
5 minutes.

It will still (attempt to) get triggered every minute -- if that fails,
the reserved concurrency of 1 will cause a "Throttled" response, which
is fine.

This way, we'll keep actual timeouts to a minimum while keeping good
response time.