fix: widen canary margin
The Inventory Canary plays clever tricks with Lambda timeouts: it
suspends itself when it detects there's only 1 minute left until
the Lambda timeout hits.

With this configuration, we hit about 1 timeout per day, indicating
that we are underestimating the time we need to finish the final
processing.

Raise the margin to 2 minutes in a bid to reduce the timeouts.