fix: transliterator breaks on packages with too many submodules (#1007)
The transliterator uploads and returns 2 files for every language for every submodule in a package.

This may exceed the maximum size of the Step Functions return value when there are very many submodules, such as a pathological case of `@dschmidt/google-provider@0.0.1` has (there are more than 600 submodules, probably because of a configuration mistake).

We could deal with this by changing the protocol (large scale changes and many tests necessary). Given that it's probably an accident, we'll just limit how many submodules we will translate.

We could hardcode to a fixed number, but that could still blow up given long enough names. So instead we'll estimate the payload size added by each submodule and use that to cap.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*