feat(transliterator): allow running `lsof` on heartbeat
Setting the `RUN_LSOF_ON_HEARTBEAT` environment variable in the task
to a non-empty valu will cause `lsof` to be run after a Step Functions
heartbeat has been initiated. This can be used to investigate open file
descriptor usage.

A simple way to ad-hoc test this is to manually create a "debug" version
of the Orchestration state machine using a copy of it's JSON definition,
changing the vlaue of `RUN_LSOF_ON_HEARTBEAT` in the container overrides
property.