fix: ingestion function tries to execute more than once if the package name and version are duplicated (#1159)
If the ingestion function receives two records for the same package, it will try to start the execution of the step function twice with the same name but different inputs, because the version ID returned by S3 is part of the input. In this case, the SDK will throw an [`ExecutionAlreadyExists`][1] error.

Prevent this by accumulating the packages that were seen in a set. Only unseen packages are processed.

[1]: https://docs.aws.amazon.com/step-functions/latest/apireference/API_StartExecution.html

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*