fix: tar-stream causes a promise deadlock
Due to a recent update in `tar-stream` (to 3.0.0), large streams passed
to `Extractor.on('entry', ...)` must be consumed before calling `next`,
or the overall tar stream will not be able to make any further progress.

Previously, we were only consuming entries that were of interest to the
extractor, and this PR changes it to read all entries, but only store
data for those of interest.

This reuqired changing how the unit tests work so that they actually
use a "real" .tar.gz data blob, as the downstream code uses `.pipe()` on
it, which requires a "real" stream (mocking it is too tedious &
error-prone).