chore: add dynamic require failure to runbook (#1315)
This PR moves the "Errors Encountered in the Past" section from the
`operator-runbook` to a new document named `errors-encountered`. An
entry for `Transliterator` task failures caused by missing files has
been added to the `errors-encountered` document which explains that the
cause may be from a dynamic require being used in a dependency. A link
to the `errors-encountered` doc has been added to the appendix section
in the `operator-runbook`.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Signed-off-by: Francis <colifran@amazon.com>