fix: stager fails when the file to download doesn't exist (#1321)
Sometimes, the stager function receives a message to download a file
that should exist but doesn' t. If we throw an error, the message will
be sent to the DLQ, to be processed again in a re-drive. But given that
this file will probably never be available, the re-drives will also
fail, and we'll never be able to get rid of the message in the DLQ.
Instead, we ignore this version by returning silently.

NPM seems to occasionally drop tarballs on the floor when a new package
version is released, which you can notice on the download stats because
the package has been downloaded `0` times. So far we've seen that happen
for:

*
[@pepperize/cdk-vpc@0.0.785](https://www.npmjs.com/package/@pepperize/cdk-vpc/v/0.0.785?activeTab=versions)
*
[@aws-sdk/credential-provider-http@3.422.0](https://www.npmjs.com/package/@aws-sdk/credential-provider-http/v/3.422.0?activeTab=versions)

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>