fix: submodule doc upload is incorrect (#1349)
When rendering and uploading documentation for the submodules, we are
using the `Assembly.submodules` property and taking only the second
component. This does not match up with the actual submodules. For
example, in the `cdk-nag` assembly, there are 46 submodules with prefix
`cdk-nag.rules`: `cdk-nag.rules`, `cdk-nag.rules.apigw`,
`cdk-nag.rules.appsync` etc. This logic creates an array with 46
repetitions of the string `rules`.

Change the selection logic to use the names provided by the
`listSubmodules()` method.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*