fix: Feed builder lambda times out (#1381)
## Problem

The feed builder lambda was timing out and consistently failing causing
the atom feed to not be updated.

## Solution

After investigation, it was determined the default memory, 128MB, was
insufficient for the lambda to correctly execute. This PR increases the
memory to 1024MB, a number verified by dev testing, and adding an alarm
such that future issues will be flagged.

### Investigation notes

Below you can see the metrics from the lambda function before and after
increasing the memory allocation, approximately 21:45UTC.

<img width="1402" alt="Screenshot 2024-01-30 at 3 35 59 PM"
src="https://github.com/cdklabs/construct-hub/assets/139287474/42e47e4e-aaef-4daa-91a1-e3f637dfd152">

Note that when the issue was resolved the invocations, throttles and
events spike, then return to a low steady state. I believe this is due
to the long backlog of work the lambda had to work through as it hasn't
run correctly in ~seven months.

This fixes #1238.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*