fix: Updating SLA-Breached evaluation periods (#1398)
## Problem

We are observing that upon resolution of npmjs replica lag, the
`SLA-Breached` alarm triggers during the period of time where the
application consumes the updates from the backlog. The alarm is self
resolving as there is no issue with the backend.

## Solution

Update the evaluation period from 2 (10 minutes) to 4 (20 minutes) to
allow for the system to catch up before reporting an alarm.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>