fix(transliterator): submodule pages are missing (#1433)
In https://github.com/cdklabs/construct-hub/pull/1007 we implemented a
check to limit the output of the transliterator task by not
transliterating all submodules of a package if the expected size is too
large to fit into a Step Functions output value. This was made on the
assumption that packages only have that many submodules by accident. As
it turns out, there are packages like `@cdktf/provider-aws` that do have
a lot of submodules by design.

The original fix assumed that the task output was required by the
Orchestration State Machine. Luckily this is not actually the case and
we can simply not return the full list of entries and in fact fully
discard the result of the Transliteration step in the Orchestration
State Machine.

This change reverts to the original, unlimited implementation. It also
keeps the function with list output around for testing purposes. However
it is now renamed and a new wrapper around it discards the potentially
oversized return data.

Fixes #1114

----

Changes to the integ test were made because the test didn't deploy for
me. I tried to simplify it a bit, to speed up deployment and needed to
add some missing permissions.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>