fix(transliterator): use sdk v3 and adaptive throttling (#1442)
Migrates the transliterator to SDKv3. With SDKv3, we can replace the
custom retry mechanism with adaptive throttling that's built-in to the
sdk.

Migration included CodeArtifact login and removal of `fs-extra` in favor
of native node features.

Also upgrades the Node.js version to node10 and the Amazon Linux version
to AL2023.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>