fix(transliterator): s3 requests are capped after 3 retries (#1445)
The previous retry strategy used with the SDK v2 did retry for 12 times
before the back-off tipped over a threshold.
Because we are seeing tasks fail due to S3 requests throttled, this PR
brings the retries for SDK v3 in line with the previous amount. It also
adds environment variables to make the s3 concurrency and retry limits
configurable via environment variables to allow for better testing.


----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*