fix(feed-builder): updating feed fails when package has no release notes (#1453)
In #1451 the update-feed function was migrated to SDK v3. For the
migration the error handling needed to be adapted to SDK v3. With SDK v2
we checked for a `404` error status to conclude that no release notes
are available. SDK v3 does not use http error codes, so we needed to
convert to error names/types. Initially this was incompletely changed to
only check for a `NotFound` error. However it also needs to check for
`NoSuchKey` error.

This has been fixed and the testcase updated.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*