fix(catalog-builder): average execution time is 100x slower than previously (#1455)
In #1441 we upgraded the SDK to v3. During this upgrade we replaced the
deprecated InvokeAsyncCommand with InvokeCommand. It was missed that the
default invocation type for this command is `RequestResponse`, when it
needs to be `Event`. This caused the catalog-builder function to await
the completion of the feed-builder function.


----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>