fix: migrate various missed code-paths to SDK v3 (#1464)
**Changes to production code**
- Migrated **inventory canary** function to SDK v3.
- Migrated **package versions table widget** function to SDK v3.
- Pulled out a new shared helper function to decompresses a possibly
gzip'ed blob stream into a string. It's used in the above migrations.
- Updated imports in some lambda handlers from `import * as aws` to
explicitly importing the used clients to make it more obvious which code
has already been migrated.
- Removed some now unused SDK v2 clients from the shared client code.
- Fixed an issue in the shared `s3ObjectExists` helper that would only
return false when the bucket is not found, but not when the key is not
found.

**Changes to test cases**
- Cleanup of old mocks in some unit tests.
- Refactor deny-list integ test to use assertions instead of triggers.
As it was, this integ test still blocked automated dependencies updates
because the hash of the trigger functions would change, causing a
replacement of the trigger function. This was classified as a
destructive change and failed the test. With assertions, these kind of
changes are excluded.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*