fix(prune-handler): upgrade to SDK v3 (#1466)
Upgrade prune-handler and prune-queue-handler to AWS SDK v3. These are
the last functions that used AWS SDK v2, so the dependency on v2 has
also been removed.

In the prune handler, there is a piece of code that is executed inside a
for loop that iterates on the deny list. It calls an "on-change"
handler, which, turns out to be the catalog builder. But there is no
need to call this function repeatedly. Only once after all the deny list
has been processed. So I moved it outside the for loop.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*