feat: expose the regenerateAllDocumentationPerPackage state machine (#1494)
Exposes the orchestration construct. This is useful if you need to use
parts of within a large scope. For example, if you need to start the
execution of the `regenerateAllDocumentationPerPackage` in the
integration step of a CI/CD pipeline:

```ts
const constructHub = new ConstructHub(/*...*/);
constructHub.regenerateAllDocumentationPerPackage.grantPrincipal.addToPrincipalPolicy(
  new PolicyStatement({
    effect: Effect.ALLOW,
    actions: ['sts:AssumeRole'],
    principals: [cicdRole],
  })
);
```

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*