fix(transliterator): service constraints cause failures to go undetected (#1506)
The configuration for the transliterator task was previously set-up to
minimize false positives and prioritize self-recovery of the system in
case of unspecified transient issues. However this approach can lead to
masking of problems when the task uses more resources then it should,
just not enough to tip it over the limit (see
https://github.com/cdklabs/construct-hub/issues/982). The approach also
causes alarms to trip off way too late and use up a lot of resources to
verify the failure state. E.g. before the recent changes, it would take
over 10hrs to alarm for a transliteration task that is consistently
timing out, with 10hrs of ECS computing time wasted on the task.

Therefore this PR is making the following adjustments:
- **Set a much stricter time out for the translator task: 15min**
Currently the task takes a maximum of ~8mins for the largest known
package. From observation, the time to process a package grows linear
with the number of snippets and sub-modules, with the latter having the
larger impact. This is because each sub-module requires the upload of
files to S3, whereas snippets are processed in memory in the same
process. There is also a base processing time that (mostly to load and
parse the jsii assembly file) that does not change much with size.
Therefore 15min will provide plenty of room for packages to grow before
reaching the limit in normal operation. However in out-of-bounds
operation, this will allow a much faster detection of failure.
- **Limit retries of all unknown errors: 2 retries with back-off**
Previously the task had two separate retry configurations for timeouts
and for any remaining errors. This mistakenly implied that timeouts are
expected and just need to be retried often enough to eventually work.
This is not correct and timeouts are never expected. We however need to
acknowledge that transient network issues might occur and aside from in
process retries will now allow for a maximum of 2 retries from the
StateMachine for any unexpected error, including time outs. In addition,
the retries will be attempted with a back-off.
- **Re-establish ulimit for NOFILE: max 1024 file descriptors at a
time**
ECS allows for configuration of ulimits. Previously we removed the
configured limit assuming that this was a misconfiguration causing the
transliteration task to take much longer. However subsequently we have
identified and fixed the actual cause for the high usage of file
descriptors. Further investigation showed that the actual expected use
of file descriptors should be in the tens (i.e. <100). Allowing the use
of so many file descriptors that should not be required, causes risk of
masking problems. I.e. with
https://github.com/cdklabs/construct-hub/issues/982 we kept increasing
the number of file descriptors assuming this would resolve the issue.
However what happened is that we just extended the timespan until the
issue would occur again. Therefore it is better to be much more
restrictive and fail faster. In a future PR we will also add runnbook
instructions on how to investigate issues with file descriptors. The
limit of 1024 was chosen match Lambda's limit. While this task is
currently running inside ECS Fargate, the expectation is that it should
also be able to run inside a Lambda Function. The limit should still be
restrictive enough to catch any bugs that inadvertently overuse the file
system, while providing a decent buffer for genuine usage increases.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*