feat: lambda function to manually re-stage a package version (#1515)
Resolves #536

Adds a new lambda function intended for manual invocation. This function
will re-attempt to stage a specific package version. Documented usage in
the runbook.

I need to upgrade the node types to node18 for this, which required some
additional type only changes w.r.t the SDKv3 usage.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*