fix: reprocess workflow fails when output is too large
StepFunctions has a limit on input or output size for a task, state, or execution of 256 KiB of data as a UTF-8 encoded string.

In the case of the reprocess workflow, if the output of `ListObjectsV2` exceeds this limit, the state machine execution fails, which interrupts the chain. As a result, only a subset of packages is reprocessed.