fix: reprocess workflow fails when output is too large (#1635)
StepFunctions has a limit on input or output size for a task, state, or
execution of 256 KiB of data as a UTF-8 encoded string.

In the case of the reprocess workflow, if the output of `ListObjectsV2`
exceeds this limit, the state machine execution fails, which interrupts
the chain. As a result, only a subset of packages is reprocessed.

Reduce the number of requested keys in each `ListObjectsV2` from 1000 to
900, to reduce the risk of this happening. With the data we have at the
moment, outputs don't exceed 200 KiB.


----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>