fix: decrement MaxKeys from 1000 by steps of 100 when encountering DataLimitExceeded (#1648)
## Description
When MaxKeys is 1000 (default), sometimes the
`ReprocessIngestionWorkflow` fails due to the `DataLimitExceeded` state
error. The limit is 256KB according to [Step Functions service
quotas](https://docs.aws.amazon.com/step-functions/latest/dg/service-quotas.html)
and cannot be reconfigured.

Reducing MaxKeys can bring us below the service quota, but increases
workflow running time. This change runs the workflow at `maxKeys: 1000`,
and reruns it with MaxKeys decremented by 100 if the workflow fails on
the `DataLimitExceeded` state error, with a floor of `maxKeys: 100`.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>