fix: discard error message when DataLimitExceeded, use original input para… (#1651)
…meters

Fixes a bug introduced by #1648, which caused the error info to
overwrite the original payload.

Example error log:  
```
An error occurred while executing the state 'S3.ListObjectsV2(NextPage)Try900' (entered at the event id #9). The JSONPath '$.ContinuationToken' specified for the field 'ContinuationToken.$' could not be found in the input '{"Error":"States.DataLimitExceeded","Cause":"The state/task 'aws-sdk:s3' returned a result with a size exceeding the maximum number of bytes service limit."}'
```

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>