chore: use new npm replica registry endpoints (#1680)
Closes #1664 

NPM recently announced a change to their replication protocol with a
hard deadline of May 29:
https://github.com/orgs/community/discussions/152515

This PR puts us in compliance with those new protocols, which we can
access via the `npm-replication-opt-in: true` header. Full details are
in the NPM announcement and subsequent discussion.

We now have to make a GET request per package to get the metadata
necessary to deal with each change. This is a departure from the
protocol before, where we were able to get this metadata alongside the
database changes. This didn't seem to cause a big change in ingestion
time, but we have parallelized the GETs anyhow.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>