feat: npm source bucket is versioned (#1693)
The source bucket stores state -- the marker sequence and the list of
known versions that have already been ingested. we are currently
resetting the state without backing the old state up, and when the new
state is malformed we have no way of rolling back.

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>