fix(npmjs follower): hangs when doing a backfill
For reasons of the NPMJS replica behaving strangely[1], our follower has
changed to starting a full backfill of the entire catalog.

However, when doing the full backfill we're seeing strange behavior:
when doing fetches for a lot of packages in parallel, we notice that the
follower stops printing anything and the Lambda just times out after 5
minutes of inactivity. Looking at the logs, 47 requests are sent out in
parallel but only 15 are answered, and then nothing more happens.

Looking at the code it is a mess of NodeJS event handling, and from what
I can tell there is no handling of socket error events that occur in the
request (just socket error events that occur in the response). Best
theory: since these events are dropped on the floor, the corresponding
Promise is never resolved or rejected and the flow will just hang.
There also was no request timeout, I added that as well.

I refactored the code to be more straightforward:

- No recursion to do retries but a `while` loop
- Give a deadline-driven timeout
- Put the classification of what makes an error retryable in once place
- Switch to using `await new Promise(...)` instead of the callback
  style.
- Importantly: add `req.on('error')` and `req.on('timeout')` to handle
  problems with the request.

[1]: https://github.com/orgs/community/discussions/152515#discussioncomment-13053455