fix(npmjs): package fetches are limited to 4 concurrent connections
When we were fetching package metadata for a lot of packages at a time
(think: 100 requests in parallel), we could observe they would get
serviced fast initially, and parallel requests later on in the batch
would get progressively slower. The last one would take on the order of
~10 seconds.

After a lot of debugging, it turns out that was because we were imposing
an artificial limit of at most 4 concurrent connections per host on
ourselves. So metadata requests 97 is waiting for requests 1-93 to
complete before it can even start.

This PR ups the concurrent connection limit also to 100, so we can
service all of them in parallel. Note: this may be slightly too much, as
I'm observing throttles from npmjs by doing this. We can play
around with this setting a little, but 4 is obviously too little.