fix(npmjs): package fetches are limited to 4 concurrent connections (#1756)
When we were fetching package metadata for a lot of packages at a time
(think: 100 requests in parallel), we could observe they would get
serviced fast initially, and parallel requests later on in the batch
would get progressively slower. The last one would take on the order of
~10 seconds.

After a lot of debugging, it turns out that was because we were imposing
an artificial limit of at most 4 concurrent connections per host on
ourselves. So metadata request 97 is waiting for requests 1-93 to
complete before it can even start.

This PR ups the concurrent connection limit also to 100, so we can
service all of them in parallel. Note: this may be slightly too much, as
I'm observing throttles from npmjs by doing this. We can play around
with this setting a little, but 4 is obviously too little.

This should improve backfill speed by a factor of ~20 (rough guess).

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>