fix: orchestration dlq alarm is too sensitive (#1758)
We are observing sporadic occurrences where the size of the DLQ is
reported as 1 for a very short period of time (less than a minute). This
triggers an alarm which is quickly auto resolved.

Its not entirely clear why this happens. There is no auto/human re-drive
when it happens. I am now attributing it to SQS eventual consistency and
deeming the alarm to sensitive.

Increase the evaluation periods to 2. This will now be consistent with
our [ingestion dlq
alarm](https://github.com/cdklabs/construct-hub/blob/main/src/backend/ingestion/index.ts#L311),
which already uses 2 evaluation periods.

> Also see https://github.com/cdklabs/construct-hub/pull/1402

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*